let's do a* search :3

+++ src/main/java/org/devathon/contest2016/intelligence/MinecraftRating.java
@@ -23,7 +23,7 @@ public class MinecraftRating extends Rating{
         return pathCosts;
     }
 
-    public void setPathCosts(float pathCosts) {
+    public void setPathCosts(int pathCosts) {
         this.pathCosts = pathCosts + RATING_PATH;
     }
